+++ src/main/java/com/podio/app/ApplicationInstall.java
@@ -8,24 +8,24 @@ public class ApplicationInstall implements Serializable {
 
 	private static final long serialVersionUID = 1L;
 
-	private int spaceId;
+	private long spaceId;
 
 	public ApplicationInstall() {
 		super();
 	}
 
-	public ApplicationInstall(int spaceId) {
+	public ApplicationInstall(long spaceId) {
 		super();
 		this.spaceId = spaceId;
 	}
 
 	@JsonProperty("space_id")
-	public int getSpaceId() {
+	public long getSpaceId() {
 		return spaceId;
 	}
 
 	@JsonProperty("space_id")
-	public void setSpaceId(int spaceId) {
+	public void setSpaceId(long spaceId) {
 		this.spaceId = spaceId;
 	}
