+++ src/main/java/com/podio/comment/CommentAPI.java
@@ -62,7 +62,7 @@ public class CommentAPI extends BaseAPI {
 	 * @param silent
 	 *            True if the update should be silent, false otherwise
 	 */
-	public int addComment(Reference reference, CommentCreate comment,
+	public long addComment(Reference reference, CommentCreate comment,
 			boolean silent, boolean hook) {
 		return getResourceFactory()
 				.getApiResource(
