+++ src/test/java/com/podio/item/ItemAPITest.java
@@ -30,7 +30,7 @@ public class ItemAPITest {
 
 	@Test
 	public void addItem() {
-		int itemId = getAPI().addItem(
+		long itemId = getAPI().addItem(
 				1,
 				new ItemCreate(null, Arrays.asList(new FieldValuesUpdate(1,
 						"value", "yes")), Collections.<Integer> emptyList(),
@@ -41,7 +41,7 @@ public class ItemAPITest {
 
 	@Test
 	public void addItemAsApp() {
-		int itemId = APIFactoryProvider
+		long itemId = APIFactoryProvider
 				.getApp(1)
 				.getAPI(ItemAPI.class)
 				.addItem(
